Add files via upload
The analysis was performed using SQL as required by the assignment. I have provided the PDF summaries of the results of specific parts of the assignment. Vaccinations.db is a combined database sourced from a sample of hand-selected spreadsheets from the source. Queries.sql has the lines of SQL commands to answer specific questions over the dataset. Design.pdf is a fully hand-crafter Entity-Relationship Diagram structuring my vision of the sampled datasets. Database.sql has the lines of SQL commands to create the Vaccinations.db file.